Git/options/requirements.txt
pandas==2.2.3
python-binance==1.0.28
python-dotenv==1.1.1
pymongo==4.13.2
pytz
requests
numpy
scipy
